Combine AlignOverview.tsx and BlastOverview.tsx into Overview.tsx.
